refactor: API 계층 분리
