+++ src/domains/restaurantAPI.ts
@@ -1,7 +1,28 @@
-import RULES from "../constants/rules.js";
-import { Restaurant } from "./restaurants.ts";
+import { Restaurant } from "./restaurant.ts";
 
 const API_URL = import.meta.env.VITE_API_URL;
+
+export async function fetchRestaurants(): Promise<Restaurant[]> {
+  try {
+    const response = await fetch(
+      `${import.meta.env.VITE_API_URL}/restaurants`,
+      {
+        method: "GET",
+        headers: { "Content-Type": "application/json" },
+      }
+    );
+
+    if (!response.ok) {
+      throw new Error(`HTTP Error! Status: ${response.status}`);
+    }
+
+    const data: Restaurant[] = await response.json();
+    return data;
+  } catch (error) {
+    alert("데이터를 불러오지 못했습니다. 다시 시도해주세요.");
+    return [];
+  }
+}
 
 export async function addRestaurant(
   restaurants: Restaurant[],
@@ -75,35 +96,4 @@ export async function toggleFavorite(
     alert("즐겨찾기를 업데이트하지 못했습니다. 다시 시도해주세요.");
     return restaurants;
   }
-}
-
-export function filterAndSortRestaurants(
-  restaurants: Restaurant[],
-  category: string,
-  sorting: string
-): Restaurant[] {
-  let filteredRestaurants = [...restaurants];
-
-  if (category !== RULES.ALL_CATEGORY) {
-    filteredRestaurants = filteredRestaurants.filter(
-      (restaurant) => restaurant.category === category
-    );
-  }
-
-  if (sorting === RULES.SORTING[1]) {
-    filteredRestaurants.sort((a, b) => {
-      const diff = a.distance - b.distance;
-      return diff !== 0 ? diff : a.name.localeCompare(b.name, "ko");
-    });
-    return filteredRestaurants;
-  }
-
-  filteredRestaurants.sort((a, b) => a.name.localeCompare(b.name, "ko"));
-  return filteredRestaurants;
-}
-
-export function getFavoriteRestaurants(
-  restaurants: Restaurant[]
-): Restaurant[] {
-  return restaurants.filter((restaurant) => restaurant.isFavorite);
 }
